third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-fling-with-page-scale-expected.txt

Tests that slow-path flings convert to the scaled coordinate space correctly. To test manually, pinch zoom into the page and perform a touchpad fling. If the mouse passes over the div it should cause it to scroll. If it doesn't pass over the div it shouldn't scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flinging vertically with mouse over container. Container should be scrolled.
PASS verticalContainer.scrollTop is not fullyScrolled
Flinging horizontally with mouse over container. Container should be scrolled.
PASS horizontalContainer.scrollLeft is not fullyScrolled
PASS successfullyParsed is true

TEST COMPLETE
